feat: 앱내의 모든 Alert를 관리하는 AlertManager 구조체 구현
